pyproject.toml
[tool.poetry]
name = "fd_client"
version = "0.1.0"
description = ""
authors = ["whatislove118 <oleg.pochikalin@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.0"
python-dotenv = "^0.20.0"
requests-toolbelt = "^0.9.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
